backend/requirements.txt
amqp==1.4.9
aniso8601==1.1.0
anyjson==0.3.3
arrow==0.8.0
billiard==3.3.0.23
celery==3.1.23
click==6.6
feedparser==5.2.1
Flask==0.11.1
Flask-RESTful==0.3.5
itsdangerous==0.24
Jinja2==2.8
kombu==3.0.35
MarkupSafe==0.23
PyJWT==1.4.2
PyMySQL==0.7.8
python-dateutil==2.5.3
pytz==2016.6.1
redis==2.10.5
requests==2.11.1
six==1.10.0
SQLAlchemy==1.0.15
Werkzeug==0.11.11
wheel==0.24.0
